New routing and controllers system Changes router to Gorilla/Mux and new controllers system with render, session and container dependency injection
