+++ src/web/controllers/backend/index.go
@@ -1,0 +1,13 @@
+package backend
+
+import (
+	sessions "github.com/goincremental/negroni-sessions"
+	"github.com/unrolled/render"
+	"net/http"
+	"web/helpers/mynegroni"
+)
+
+func Index(rw http.ResponseWriter, r *http.Request, render *render.Render, s sessions.Session, c *mynegroni.Content) {
+
+	render.HTML(rw, http.StatusOK, "backend/index", c)
+}
